Added tesseract test case
